Organise user controller files
